+++ tests/actor/_power_managing/test_matryoshka.py
@@ -12,6 +12,7 @@ from frequenz.quantities import Power
 
 from frequenz.sdk import timeseries
 from frequenz.sdk.microgrid._power_managing import Proposal
+from frequenz.sdk.microgrid._power_managing._base_classes import DefaultPower
 from frequenz.sdk.microgrid._power_managing._matryoshka import Matryoshka
 from frequenz.sdk.timeseries import _base_types
 
@@ -28,7 +29,9 @@ class StatefulTester:
         self._call_count = 0
         self._batteries = batteries
         self._system_bounds = system_bounds
-        self.algorithm = Matryoshka(max_proposal_age=timedelta(seconds=60.0))
+        self.algorithm = Matryoshka(
+            max_proposal_age=timedelta(seconds=60.0), default_power=DefaultPower.ZERO
+        )
 
     def tgt_power(  # pylint: disable=too-many-arguments,too-many-positional-arguments
         self,
@@ -37,7 +40,6 @@ class StatefulTester:
         bounds: tuple[float | None, float | None],
         expected: float | None,
         creation_time: float | None = None,
-        must_send: bool = False,
         batteries: frozenset[int] | None = None,
     ) -> None:
         """Test the target power calculation."""
@@ -60,7 +62,6 @@ class StatefulTester:
                 ),
             ),
             self._system_bounds,
-            must_send,
         )
         assert tgt_power == (
             Power.from_watts(expected) if expected is not None else None
@@ -109,10 +110,8 @@ async def test_matryoshka_no_excl() -> None:  # pylint: disable=too-many-stateme
     tester.bounds(priority=2, expected_power=25.0, expected_bounds=(-200.0, 200.0))
     tester.bounds(priority=1, expected_power=25.0, expected_bounds=(25.0, 50.0))
 
-    tester.tgt_power(priority=1, power=20.0, bounds=(20.0, 50.0), expected=None)
-    tester.tgt_power(
-        priority=1, power=20.0, bounds=(20.0, 50.0), expected=25.0, must_send=True
-    )
+    tester.tgt_power(priority=1, power=20.0, bounds=(20.0, 50.0), expected=25.0)
+    tester.tgt_power(priority=1, power=20.0, bounds=(20.0, 50.0), expected=25.0)
     tester.bounds(priority=1, expected_power=25.0, expected_bounds=(25.0, 50.0))
 
     tester.tgt_power(priority=3, power=10.0, bounds=(10.0, 15.0), expected=15.0)
@@ -125,7 +124,7 @@ async def test_matryoshka_no_excl() -> None:  # pylint: disable=too-many-stateme
     tester.bounds(priority=2, expected_power=22.0, expected_bounds=(10.0, 22.0))
     tester.bounds(priority=1, expected_power=22.0, expected_bounds=(10.0, 22.0))
 
-    tester.tgt_power(priority=1, power=30.0, bounds=(20.0, 50.0), expected=None)
+    tester.tgt_power(priority=1, power=30.0, bounds=(20.0, 50.0), expected=22.0)
     tester.bounds(priority=1, expected_power=22.0, expected_bounds=(10.0, 22.0))
 
     tester.tgt_power(priority=3, power=10.0, bounds=(10.0, 50.0), expected=30.0)
@@ -150,14 +149,14 @@ async def test_matryoshka_no_excl() -> None:  # pylint: disable=too-many-stateme
     tester.bounds(priority=2, expected_power=-50.0, expected_bounds=(-200.0, -50.0))
     tester.bounds(priority=1, expected_power=-50.0, expected_bounds=(-200.0, -50.0))
 
-    tester.tgt_power(priority=3, power=-0.0, bounds=(-200.0, 200.0), expected=None)
+    tester.tgt_power(priority=3, power=-0.0, bounds=(-200.0, 200.0), expected=-50.0)
     tester.bounds(priority=1, expected_power=-50.0, expected_bounds=(-200.0, -50.0))
 
     tester.tgt_power(priority=1, power=-150.0, bounds=(-200.0, -150.0), expected=-150.0)
     tester.bounds(priority=2, expected_power=-150.0, expected_bounds=(-200.0, -50.0))
     tester.bounds(priority=1, expected_power=-150.0, expected_bounds=(-200.0, -50.0))
 
-    tester.tgt_power(priority=4, power=-180.0, bounds=(-200.0, -50.0), expected=None)
+    tester.tgt_power(priority=4, power=-180.0, bounds=(-200.0, -50.0), expected=-150.0)
     tester.bounds(priority=1, expected_power=-150.0, expected_bounds=(-200.0, -50.0))
 
     tester.tgt_power(priority=4, power=50.0, bounds=(50.0, None), expected=50.0)
@@ -172,7 +171,7 @@ async def test_matryoshka_no_excl() -> None:  # pylint: disable=too-many-stateme
     tester.bounds(priority=2, expected_power=-150.0, expected_bounds=(-200.0, 200.0))
     tester.bounds(priority=1, expected_power=-150.0, expected_bounds=(-200.0, 200.0))
 
-    tester.tgt_power(priority=3, power=0.0, bounds=(-200.0, 200.0), expected=None)
+    tester.tgt_power(priority=3, power=0.0, bounds=(-200.0, 200.0), expected=-150.0)
     tester.bounds(priority=3, expected_power=-150.0, expected_bounds=(-200.0, 200.0))
     tester.bounds(priority=2, expected_power=-150.0, expected_bounds=(-200.0, 200.0))
     tester.bounds(priority=1, expected_power=-150.0, expected_bounds=(-200.0, 200.0))
@@ -221,7 +220,7 @@ async def test_matryoshka_with_excl_1() -> None:
     tester.bounds(priority=2, expected_power=25.0, expected_bounds=(-200.0, 200.0))
     tester.bounds(priority=1, expected_power=25.0, expected_bounds=(25.0, 50.0))
 
-    tester.tgt_power(priority=1, power=20.0, bounds=(20.0, 50.0), expected=None)
+    tester.tgt_power(priority=1, power=20.0, bounds=(20.0, 50.0), expected=25.0)
     tester.bounds(priority=1, expected_power=25.0, expected_bounds=(25.0, 50.0))
 
     tester.tgt_power(priority=2, power=-10.0, bounds=(-10.0, 50.0), expected=20.0)
@@ -231,17 +230,17 @@ async def test_matryoshka_with_excl_1() -> None:
     tester.tgt_power(priority=1, power=-10.0, bounds=(-10.0, 50.0), expected=0.0)
     tester.bounds(priority=0, expected_power=0.0, expected_bounds=(0.0, 50.0))
 
-    tester.tgt_power(priority=1, power=-10.0, bounds=(-10.0, 20.0), expected=None)
+    tester.tgt_power(priority=1, power=-10.0, bounds=(-10.0, 20.0), expected=0.0)
     tester.bounds(priority=0, expected_power=0.0, expected_bounds=(0.0, 20.0))
 
-    tester.tgt_power(priority=1, power=-10.0, bounds=(-10.0, -5.0), expected=None)
+    tester.tgt_power(priority=1, power=-10.0, bounds=(-10.0, -5.0), expected=0.0)
     tester.bounds(priority=0, expected_power=0.0, expected_bounds=(0.0, 50.0))
 
     tester.tgt_power(priority=2, power=-10.0, bounds=(-200.0, -5.0), expected=-30.0)
     tester.bounds(priority=1, expected_power=-30.0, expected_bounds=(-200.0, -30.0))
     tester.bounds(priority=0, expected_power=-30.0, expected_bounds=(-200.0, -30.0))
 
-    tester.tgt_power(priority=1, power=-10.0, bounds=(-100.0, -5.0), expected=None)
+    tester.tgt_power(priority=1, power=-10.0, bounds=(-100.0, -5.0), expected=-30.0)
     tester.bounds(priority=0, expected_power=-30.0, expected_bounds=(-100.0, -30.0))
 
     tester.tgt_power(priority=1, power=-40.0, bounds=(-100.0, -35.0), expected=-40.0)
@@ -271,10 +270,10 @@ async def test_matryoshka_with_excl_2() -> None:
     tester.bounds(priority=2, expected_power=30.0, expected_bounds=(-200.0, 200.0))
     tester.bounds(priority=1, expected_power=30.0, expected_bounds=(30.0, 50.0))
 
-    tester.tgt_power(priority=1, power=20.0, bounds=(20.0, 50.0), expected=None)
+    tester.tgt_power(priority=1, power=20.0, bounds=(20.0, 50.0), expected=30.0)
     tester.bounds(priority=1, expected_power=30.0, expected_bounds=(30.0, 50.0))
 
-    tester.tgt_power(priority=1, power=10.0, bounds=(5.0, 10.0), expected=None)
+    tester.tgt_power(priority=1, power=10.0, bounds=(5.0, 10.0), expected=30.0)
     tester.bounds(priority=0, expected_power=30.0, expected_bounds=(30, 50.0))
 
     tester.tgt_power(priority=2, power=-10.0, bounds=(-10.0, 50.0), expected=0.0)
@@ -290,17 +289,17 @@ async def test_matryoshka_with_excl_2() -> None:
     tester.tgt_power(priority=1, power=-10.0, bounds=(-10.0, 50.0), expected=-10.0)
     tester.bounds(priority=0, expected_power=-10.0, expected_bounds=(-10.0, 50.0))
 
-    tester.tgt_power(priority=1, power=-10.0, bounds=(-10.0, 20.0), expected=None)
+    tester.tgt_power(priority=1, power=-10.0, bounds=(-10.0, 20.0), expected=-10.0)
     tester.bounds(priority=0, expected_power=-10.0, expected_bounds=(-10.0, 0.0))
 
-    tester.tgt_power(priority=1, power=-10.0, bounds=(-10.0, -5.0), expected=None)
+    tester.tgt_power(priority=1, power=-10.0, bounds=(-10.0, -5.0), expected=-10.0)
     tester.bounds(priority=0, expected_power=-10.0, expected_bounds=(-10.0, -5.0))
 
-    tester.tgt_power(priority=2, power=-10.0, bounds=(-200.0, -5.0), expected=None)
+    tester.tgt_power(priority=2, power=-10.0, bounds=(-200.0, -5.0), expected=-10.0)
     tester.bounds(priority=1, expected_power=-10.0, expected_bounds=(-200.0, -5.0))
     tester.bounds(priority=0, expected_power=-10.0, expected_bounds=(-10.0, -5.0))
 
-    tester.tgt_power(priority=1, power=-10.0, bounds=(-100.0, -5.0), expected=None)
+    tester.tgt_power(priority=1, power=-10.0, bounds=(-100.0, -5.0), expected=-10.0)
     tester.bounds(priority=0, expected_power=-10.0, expected_bounds=(-100.0, -5.0))
 
     tester.tgt_power(priority=1, power=-40.0, bounds=(-100.0, -35.0), expected=-40.0)
@@ -328,22 +327,22 @@ async def test_matryoshka_with_excl_3() -> None:
     tester.tgt_power(priority=2, power=10.0, bounds=(None, None), expected=30.0)
     tester.tgt_power(priority=2, power=-10.0, bounds=(None, None), expected=-30.0)
     tester.tgt_power(priority=2, power=0.0, bounds=(None, None), expected=0.0)
-    tester.tgt_power(priority=3, power=20.0, bounds=(None, None), expected=None)
+    tester.tgt_power(priority=3, power=20.0, bounds=(None, None), expected=0.0)
     tester.tgt_power(priority=1, power=-20.0, bounds=(None, None), expected=-30.0)
-    tester.tgt_power(priority=3, power=None, bounds=(None, None), expected=None)
+    tester.tgt_power(priority=3, power=None, bounds=(None, None), expected=-30.0)
     tester.tgt_power(priority=1, power=None, bounds=(None, None), expected=0.0)
 
     tester.tgt_power(priority=2, power=25.0, bounds=(25.0, 50.0), expected=30.0)
     tester.bounds(priority=2, expected_power=30.0, expected_bounds=(-200.0, 200.0))
     tester.bounds(priority=1, expected_power=30.0, expected_bounds=(30.0, 50.0))
 
-    tester.tgt_power(priority=1, power=20.0, bounds=(20.0, 50.0), expected=None)
+    tester.tgt_power(priority=1, power=20.0, bounds=(20.0, 50.0), expected=30.0)
     tester.bounds(priority=1, expected_power=30.0, expected_bounds=(30.0, 50.0))
 
-    tester.tgt_power(priority=1, power=10.0, bounds=(5.0, 10.0), expected=None)
+    tester.tgt_power(priority=1, power=10.0, bounds=(5.0, 10.0), expected=30.0)
     tester.bounds(priority=0, expected_power=30.0, expected_bounds=(30, 50.0))
 
-    tester.tgt_power(priority=2, power=-10.0, bounds=(-10.0, 50.0), expected=None)
+    tester.tgt_power(priority=2, power=-10.0, bounds=(-10.0, 50.0), expected=30.0)
     tester.bounds(priority=1, expected_power=30.0, expected_bounds=(30.0, 50.0))
     tester.bounds(priority=0, expected_power=30.0, expected_bounds=(30.0, 50.0))
 
@@ -357,7 +356,7 @@ async def test_matryoshka_with_excl_3() -> None:
     tester.bounds(priority=1, expected_power=-30.0, expected_bounds=(-200.0, -30.0))
     tester.bounds(priority=0, expected_power=-30.0, expected_bounds=(-200.0, -30.0))
 
-    tester.tgt_power(priority=1, power=-10.0, bounds=(-100.0, -5.0), expected=None)
+    tester.tgt_power(priority=1, power=-10.0, bounds=(-100.0, -5.0), expected=-30.0)
     tester.bounds(priority=0, expected_power=-30.0, expected_bounds=(-100.0, -30.0))
 
     tester.tgt_power(priority=1, power=-40.0, bounds=(-100.0, -35.0), expected=-40.0)
@@ -395,13 +394,9 @@ async def test_matryoshka_drop_old_proposals() -> None:
         expected=25.0,
     )
 
-    tester.tgt_power(
-        priority=1, power=20.0, bounds=(20.0, 50.0), expected=25.0, must_send=True
-    )
+    tester.tgt_power(priority=1, power=20.0, bounds=(20.0, 50.0), expected=25.0)
     tester.algorithm.drop_old_proposals(now)
-    tester.tgt_power(
-        priority=1, power=20.0, bounds=(20.0, 50.0), expected=22.0, must_send=True
-    )
+    tester.tgt_power(priority=1, power=20.0, bounds=(20.0, 50.0), expected=22.0)
 
     # When overwritten by a newer proposal, that proposal is not dropped.
     tester.tgt_power(
@@ -417,16 +412,11 @@ async def test_matryoshka_drop_old_proposals() -> None:
         bounds=(25.0, 50.0),
         creation_time=now - 30.0,
         expected=25.0,
-        must_send=True,
     )
 
-    tester.tgt_power(
-        priority=1, power=20.0, bounds=(20.0, 50.0), expected=25.0, must_send=True
-    )
+    tester.tgt_power(priority=1, power=20.0, bounds=(20.0, 50.0), expected=25.0)
     tester.algorithm.drop_old_proposals(now)
-    tester.tgt_power(
-        priority=1, power=20.0, bounds=(20.0, 50.0), expected=25.0, must_send=True
-    )
+    tester.tgt_power(priority=1, power=20.0, bounds=(20.0, 50.0), expected=25.0)
 
     # When all proposals are too old, they are dropped, and the buckets are dropped as
     # well.  After that, sending a request for a different but overlapping bucket will
@@ -443,7 +433,6 @@ async def test_matryoshka_drop_old_proposals() -> None:
             power=25.0,
             bounds=(25.0, 50.0),
             expected=25.0,
-            must_send=True,
             batteries=overlapping_batteries,
         )
 
@@ -453,7 +442,6 @@ async def test_matryoshka_drop_old_proposals() -> None:
         bounds=(25.0, 50.0),
         creation_time=now - 70.0,
         expected=25.0,
-        must_send=True,
     )
     tester.tgt_power(
         priority=2,
@@ -461,7 +449,6 @@ async def test_matryoshka_drop_old_proposals() -> None:
         bounds=(25.0, 50.0),
         creation_time=now - 70.0,
         expected=25.0,
-        must_send=True,
     )
     tester.tgt_power(
         priority=3,
@@ -469,7 +456,6 @@ async def test_matryoshka_drop_old_proposals() -> None:
         bounds=(25.0, 50.0),
         creation_time=now - 70.0,
         expected=25.0,
-        must_send=True,
     )
 
     tester.algorithm.drop_old_proposals(now)
@@ -479,7 +465,6 @@ async def test_matryoshka_drop_old_proposals() -> None:
         power=25.0,
         bounds=(25.0, 50.0),
         expected=25.0,
-        must_send=True,
         batteries=overlapping_batteries,
     )
 
@@ -514,7 +499,6 @@ async def test_matryoshka_none_proposals() -> None:
                 power=None,
                 bounds=(20.0, 50.0),
                 expected=None,
-                must_send=True,
                 batteries=overlapping_batteries,
             )
 
@@ -522,14 +506,14 @@ async def test_matryoshka_none_proposals() -> None:
 
     tester.tgt_power(priority=3, power=22.0, bounds=(22.0, 30.0), expected=22.0)
     tester.tgt_power(priority=2, power=25.0, bounds=(25.0, 50.0), expected=25.0)
-    tester.tgt_power(priority=1, power=20.0, bounds=(20.0, 50.0), expected=None)
+    tester.tgt_power(priority=1, power=20.0, bounds=(20.0, 50.0), expected=25.0)
 
     ensure_overlapping_bucket_request_fails()
-    tester.tgt_power(priority=1, power=None, bounds=(None, None), expected=None)
+    tester.tgt_power(priority=1, power=None, bounds=(None, None), expected=25.0)
     ensure_overlapping_bucket_request_fails()
-    tester.tgt_power(priority=3, power=None, bounds=(None, None), expected=None)
+    tester.tgt_power(priority=3, power=None, bounds=(None, None), expected=25.0)
     ensure_overlapping_bucket_request_fails()
-    tester.tgt_power(priority=2, power=None, bounds=(None, None), expected=None)
+    tester.tgt_power(priority=2, power=None, bounds=(None, None), expected=0.0)
 
     # Overlapping battery bucket is dropped.
     tester.tgt_power(