+++ tests/timeseries/_pv_pool/test_pv_pool_control_methods.py
@@ -95,7 +95,7 @@ class TestPVPoolControl:
 
     def _assert_report(  # pylint: disable=too-many-arguments
         self,
-        report: PVPoolReport,
+        report: PVPoolReport | None,
         *,
         power: float | None,
         lower: float,
@@ -105,7 +105,7 @@ class TestPVPoolControl:
             typing.Callable[[_power_distributing.Result], bool] | None
         ) = None,
     ) -> None:
-        assert report.target_power == (
+        assert report is not None and report.target_power == (
             Power.from_watts(power) if power is not None else None
         )
         assert report.bounds is not None
@@ -119,15 +119,18 @@ class TestPVPoolControl:
         self,
         bounds_rx: Receiver[PVPoolReport],
         check: typing.Callable[[PVPoolReport], bool],
-    ) -> None:
+    ) -> PVPoolReport | None:
         """Receive reports until the given condition is met."""
         max_reports = 10
         ctr = 0
+        latest_report: PVPoolReport | None = None
         while ctr < max_reports:
             ctr += 1
-            report = await bounds_rx.receive()
-            if check(report):
+            latest_report = await bounds_rx.receive()
+            if check(latest_report):
                 break
+
+        return latest_report
 
     async def test_setting_power(  # pylint: disable=too-many-statements
         self,
@@ -142,13 +145,11 @@ class TestPVPoolControl:
         await self._init_pv_inverters(mocks)
         pv_pool = microgrid.new_pv_pool(priority=5)
         bounds_rx = pv_pool.power_status.new_receiver()
-        await self._recv_reports_until(
+        latest_report = await self._recv_reports_until(
             bounds_rx,
             lambda x: x.bounds is not None and x.bounds.lower.as_watts() == -100000.0,
         )
-        self._assert_report(
-            await bounds_rx.receive(), power=None, lower=-100000.0, upper=0.0
-        )
+        self._assert_report(latest_report, power=None, lower=-100000.0, upper=0.0)
         await pv_pool.propose_power(Power.from_watts(-80000.0))
         await self._recv_reports_until(
             bounds_rx,
@@ -277,4 +278,24 @@ class TestPVPoolControl:
             mocker.call(inv_ids[1], 0.0),
             mocker.call(inv_ids[2], 0.0),
             mocker.call(inv_ids[3], 0.0),
+        ]
+
+        # Resetting the power should lead to default (full) power getting set for all
+        # inverters.
+        set_power.reset_mock()
+        await pv_pool.propose_power(None)
+        report = await self._recv_reports_until(
+            bounds_rx,
+            lambda x: x.target_power is None,
+        )
+        self._assert_report(report, power=None, lower=-100000.0, upper=0.0)
+        await asyncio.sleep(0.0)
+
+        assert set_power.call_count == 4
+        inv_ids = mocks.microgrid.pv_inverter_ids
+        assert sorted(set_power.call_args_list, key=lambda x: x.args[0]) == [
+            mocker.call(inv_ids[0], -10_000.0),
+            mocker.call(inv_ids[1], -20_000.0),
+            mocker.call(inv_ids[2], -30_000.0),
+            mocker.call(inv_ids[3], -40_000.0),
         ]