+++ RELEASE_NOTES.md
@@ -14,4 +14,11 @@
 
 ## Bug Fixes
 
-<!-- Here goes notable bug fixes that are worth a special mention or explanation -->
+- Components used to be just forgotten by the power manager when all proposals are withdrawn, leaving them at their last set power values.  This has been fixed by getting the power manager to set the components to their default powers, based on the component category (according to the table below), as the last step.
+
+
+  | component category | default power                             |
+  |--------------------|-------------------------------------------|
+  | Battery            | 0.0                                       |
+  | PV                 | Minimum power (aka max production power)  |
+  | EV Chargers        | Maximum power (aka max consumption power) |